# docs/swagger.yaml
basePath: /
definitions:
...
  domain.MAINTENANCE_LOG_STATUS:
    enum:
    - CASE_OPEN
    - PENDING
    - APPROVED
    - REJECTED
    type: string
    x-enum-varnames:
    - MAINTENANCE_LOG_STATUS_CASE_OPEN
    - MAINTENANCE_LOG_STATUS_PENDING
    - MAINTENANCE_LOG_STATUS_APPROVED
    - MAINTENANCE_LOG_STATUS_REJECTED
...
  domain.ROOM:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      roomNumber:
        type: string
      roomType:
        $ref: '#/definitions/domain.RoomType'
      updateAt:
        type: string
    type: object
  domain.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.RoomType:
    properties:
      accommodate:
        type: integer
      createdAt:
        type: string
      detail:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      room:
        items:
          $ref: '#/definitions/domain.ROOM'
        type: array
      updateAt:
        type: string
    type: object
...
  payload.CreateReservationDTO:
    properties:
      endDate:
        type: string
      paymentInfoId:
        type: integer
      price:
        type: number
      roomId:
        type: integer
      serviceId:
        type: integer
      startDate:
        type: string
    required:
    - endDate
    - paymentInfoId
    - price
    - roomId
    - serviceId
    - startDate
    type: object
...
  payload.GetRoomAvailableGroupByRoomTypeDTO:
    properties:
      endDate:
        type: string
      startDate:
        type: string
    required:
    - endDate
    - startDate
    type: object
  payload.Maintenance_Log_On_Maintenance_Create:
    properties:
      description:
        type: string
      status:
        $ref: '#/definitions/domain.MAINTENANCE_LOG_STATUS'
    required:
    - description
    - status
    type: object
  payload.MaintenanceCreateDTO:
    properties:
      maintenanceLog:
        items:
          $ref: '#/definitions/payload.Maintenance_Log_On_Maintenance_Create'
        type: array
      roomId:
        type: integer
    required:
    - roomId
    type: object
...
  payload.RoomUpdateRoomIsActiveDTO:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  title: CPE241 Project Backend API
  version: "1.0"
paths:
...
  /api/maintenance:
    post:
      consumes:
      - application/json
      description: Create Maintenance
      parameters:
      - description: Session ID
        in: header
        name: ssid
        required: true
        type: string
      - description: MAINTENANCE_CREATE
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.MaintenanceCreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Maintenance
      tags:
      - Maintenance
  /api/maintenance/{id}:
    get:
      consumes:
      - application/json
      description: GetByID Maintenance
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: GetByID Maintenance
      tags:
      - Maintenance
...
  /api/reservation:
    post:
      consumes:
      - application/json
      description: Create reservation
      parameters:
      - description: Create reservation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.CreateReservationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Create reservation
      tags:
      - reservation
  /api/reservation/get-room-available:
    post:
      consumes:
      - application/json
      description: Get room available group by room type
      parameters:
      - description: Get room available group by room type
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.GetRoomAvailableGroupByRoomTypeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get room available group by room type
      tags:
      - reservation
...
  /api/room:
    post:
      consumes:
      - application/json
      description: Create room
      parameters:
      - description: Room
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/domain.ROOM'
      produces:
      - application/json
      responses: {}
      summary: Create room
      tags:
      - room
  /api/room/{id}:
    get:
      consumes:
      - application/json
      description: Get room by id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get room by id
      tags:
      - room
    put:
      consumes:
      - application/json
      description: Update room
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Room
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/domain.ROOM'
      produces:
      - application/json
      responses: {}
      summary: Update room
      tags:
      - room
  /api/room/active:
    post:
      consumes:
      - application/json
      description: Update room is active
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Is Active
        in: body
        name: isActive
        required: true
        schema:
          $ref: '#/definitions/payload.RoomUpdateRoomIsActiveDTO'
      produces:
      - application/json
      responses: {}
      summary: Update room is active
      tags:
      - room
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"